feat: added extra endpoint to see who reposted that post

+++ backend/src/main/java/com/example/backend/services/RepostService.java
@@ -10,6 +10,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 
 @Service
@@ -30,6 +31,11 @@ public class RepostService {
         Post post = postRepository.findById(postId)
                 .orElseThrow(() -> new RuntimeException("Post not found"));
 
+        Optional<Repost> existingRepost = repostRepository.findByIdAndUserId(postId, userId);
+        if (existingRepost.isPresent()) {
+            throw new RuntimeException("User has already reposted this post");
+        }
+
         Repost repost = new Repost();
         repost.setUser(user);
         repost.setPost(post);
@@ -39,6 +45,15 @@ public class RepostService {
         postRepository.save(post);
     }
 
+
+    public List<User> getUsersWhoRepostedPost(Long postId) {
+        return repostRepository.findUsersByPostId(postId);
+    }
+
+
+
+
+
     public List<Repost> getAllRepostsByUser(Long userId) {
         return repostRepository.findByUserId(userId);
     }
@@ -47,11 +62,16 @@ public class RepostService {
         Repost repost = repostRepository.findByIdAndUserId(repostId, userId)
                 .orElseThrow(() -> new RuntimeException("Repost not found or does not belong to the user"));
 
-
         Post post = repost.getPost();
+
+        if (post == null) {
+            throw new RuntimeException("Post not found for the repost");
+        }
+
         post.setRepostsCount(post.getRepostsCount() - 1);
         postRepository.save(post);
 
         repostRepository.delete(repost);
     }
+
 }
